chore: Move KeylessBackupOrigin to types file
