+++ src/keylessBackup/KeylessBackupIntro.test.tsx
@@ -3,8 +3,7 @@ import React from 'react'
 import { KeylessBackupEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import KeylessBackupIntro from 'src/keylessBackup/KeylessBackupIntro'
-import { KeylessBackupOrigin } from 'src/keylessBackup/KeylessBackupProgress'
-import { KeylessBackupFlow } from 'src/keylessBackup/types'
+import { KeylessBackupFlow, KeylessBackupOrigin } from 'src/keylessBackup/types'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import MockedNavigator from 'test/MockedNavigator'
