+++ src/keylessBackup/KeylessBackupIntro.tsx
@@ -10,9 +10,8 @@ import CancelButton from 'src/components/CancelButton'
 import Card from 'src/components/Card'
 import TextButton from 'src/components/TextButton'
 import EnvelopeIcon from 'src/keylessBackup/EnvelopeIcon'
-import { KeylessBackupOrigin } from 'src/keylessBackup/KeylessBackupProgress'
 import SmartphoneIcon from 'src/keylessBackup/SmartphoneIcon'
-import { KeylessBackupFlow } from 'src/keylessBackup/types'
+import { KeylessBackupFlow, KeylessBackupOrigin } from 'src/keylessBackup/types'
 import { emptyHeader } from 'src/navigator/Headers'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
