+++ src/onboarding/registration/ImportSelect.test.tsx
@@ -2,8 +2,7 @@ import { fireEvent, render } from '@testing-library/react-native'
 import * as React from 'react'
 import 'react-native'
 import { Provider } from 'react-redux'
-import { KeylessBackupOrigin } from 'src/keylessBackup/KeylessBackupProgress'
-import { KeylessBackupFlow } from 'src/keylessBackup/types'
+import { KeylessBackupFlow, KeylessBackupOrigin } from 'src/keylessBackup/types'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import ImportSelect from 'src/onboarding/registration/ImportSelect'
