+++ src/onboarding/registration/ImportSelect.tsx
@@ -11,8 +11,7 @@ import Card from 'src/components/Card'
 import Touchable from 'src/components/Touchable'
 import CloudCheck from 'src/icons/CloudCheck'
 import Lock from 'src/icons/Lock'
-import { KeylessBackupOrigin } from 'src/keylessBackup/KeylessBackupProgress'
-import { KeylessBackupFlow } from 'src/keylessBackup/types'
+import { KeylessBackupFlow, KeylessBackupOrigin } from 'src/keylessBackup/types'
 import { nuxNavigationOptions } from 'src/navigator/Headers'
 import { navigate, navigateClearingStack } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
